Moved all old exceptions to logic package
 Changes to be committed:
	modified:   src/main/java/com/ardaslegends/domain/Faction.java
	modified:   src/main/java/com/ardaslegends/domain/Player.java
	modified:   src/main/java/com/ardaslegends/domain/applications/AbstractApplication.java
	modified:   src/main/java/com/ardaslegends/domain/applications/RoleplayApplication.java
	modified:   src/main/java/com/ardaslegends/domain/war/War.java
	modified:   src/main/java/com/ardaslegends/presentation/discord/commands/bind/BindArmyOrCompanyCommand.java
	modified:   src/main/java/com/ardaslegends/service/ArmyService.java
	modified:   src/main/java/com/ardaslegends/service/ClaimBuildService.java
	modified:   src/main/java/com/ardaslegends/service/FactionService.java
	modified:   src/main/java/com/ardaslegends/service/MovementService.java
	modified:   src/main/java/com/ardaslegends/service/Pathfinder.java
	modified:   src/main/java/com/ardaslegends/service/PlayerService.java
	modified:   src/main/java/com/ardaslegends/service/RegionService.java
	modified:   src/main/java/com/ardaslegends/service/RpCharService.java
	modified:   src/main/java/com/ardaslegends/service/ScheduleService.java
	modified:   src/main/java/com/ardaslegends/service/UnitTypeService.java
	modified:   src/main/java/com/ardaslegends/service/applications/ClaimbuildApplicationService.java
	modified:   src/main/java/com/ardaslegends/service/applications/RoleplayApplicationService.java
	renamed:    src/main/java/com/ardaslegends/service/exceptions/applications/ClaimbuildApplicationException.java -> src/main/java/com/ardaslegends/service/exceptions/logic/applications/ClaimbuildApplicationException.java
	renamed:    src/main/java/com/ardaslegends/service/exceptions/applications/RoleplayApplicationServiceException.java -> src/main/java/com/ardaslegends/service/exceptions/logic/applications/RoleplayApplicationServiceException.java
	renamed:    src/main/java/com/ardaslegends/service/exceptions/army/ArmyServiceException.java -> src/main/java/com/ardaslegends/service/exceptions/logic/army/ArmyServiceException.java
	renamed:    src/main/java/com/ardaslegends/service/exceptions/claimbuild/ClaimBuildServiceException.java -> src/main/java/com/ardaslegends/service/exceptions/logic/claimbuild/ClaimBuildServiceException.java
	renamed:    src/main/java/com/ardaslegends/service/exceptions/FactionServiceException.java -> src/main/java/com/ardaslegends/service/exceptions/logic/faction/FactionServiceException.java
	renamed:    src/main/java/com/ardaslegends/service/exceptions/movement/MovementServiceException.java -> src/main/java/com/ardaslegends/service/exceptions/logic/movement/MovementServiceException.java
	renamed:    src/main/java/com/ardaslegends/service/exceptions/PathfinderServiceException.java -> src/main/java/com/ardaslegends/service/exceptions/logic/movement/PathfinderServiceException.java
	renamed:    src/main/java/com/ardaslegends/service/exceptions/PlayerServiceException.java -> src/main/java/com/ardaslegends/service/exceptions/logic/player/PlayerServiceException.java
	renamed:    src/main/java/com/ardaslegends/service/exceptions/region/RegionServiceException.java -> src/main/java/com/ardaslegends/service/exceptions/logic/region/RegionServiceException.java
	renamed:    src/main/java/com/ardaslegends/service/exceptions/rpchar/RpCharServiceException.java -> src/main/java/com/ardaslegends/service/exceptions/logic/rpchar/RpCharServiceException.java
	renamed:    src/main/java/com/ardaslegends/service/exceptions/units/UnitServiceException.java -> src/main/java/com/ardaslegends/service/exceptions/logic/units/UnitServiceException.java
	renamed:    src/main/java/com/ardaslegends/service/exceptions/BattleServiceException.java -> src/main/java/com/ardaslegends/service/exceptions/logic/war/BattleServiceException.java
	renamed:    src/main/java/com/ardaslegends/service/exceptions/WarServiceException.java -> src/main/java/com/ardaslegends/service/exceptions/logic/war/WarServiceException.java
	modified:   src/main/java/com/ardaslegends/service/utils/ServiceUtils.java
	modified:   src/main/java/com/ardaslegends/service/war/BattleService.java
	modified:   src/main/java/com/ardaslegends/service/war/WarService.java
	modified:   src/test/java/com/ardaslegends/domain/FactionTest.java
	modified:   src/test/java/com/ardaslegends/service/ArmyServiceTest.java
	modified:   src/test/java/com/ardaslegends/service/ClaimbuildServiceTest.java
	modified:   src/test/java/com/ardaslegends/service/FactionServiceTest.java
	modified:   src/test/java/com/ardaslegends/service/MovementServiceTest.java
	modified:   src/test/java/com/ardaslegends/service/PathfinderTest.java
	modified:   src/test/java/com/ardaslegends/service/PlayerServiceTest.java
	modified:   src/test/java/com/ardaslegends/service/utils/ServiceUtilsTest.java
