+++ src/main/java/com/ardaslegends/service/ArmyService.java
@@ -7,10 +7,10 @@ import com.ardaslegends.repository.faction.FactionRepository;
 import com.ardaslegends.repository.MovementRepository;
 import com.ardaslegends.service.dto.army.*;
 import com.ardaslegends.service.dto.unit.UnitTypeDto;
-import com.ardaslegends.service.exceptions.FactionServiceException;
-import com.ardaslegends.service.exceptions.PlayerServiceException;
-import com.ardaslegends.service.exceptions.army.ArmyServiceException;
-import com.ardaslegends.service.exceptions.claimbuild.ClaimBuildServiceException;
+import com.ardaslegends.service.exceptions.logic.faction.FactionServiceException;
+import com.ardaslegends.service.exceptions.logic.player.PlayerServiceException;
+import com.ardaslegends.service.exceptions.logic.army.ArmyServiceException;
+import com.ardaslegends.service.exceptions.logic.claimbuild.ClaimBuildServiceException;
 import com.ardaslegends.service.utils.ServiceUtils;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
