+++ src/main/java/com/ardaslegends/service/RpCharService.java
@@ -1,10 +1,8 @@
 package com.ardaslegends.service;
 
-import com.ardaslegends.domain.ClaimBuild;
 import com.ardaslegends.domain.RPChar;
 import com.ardaslegends.repository.rpchar.RpcharRepository;
-import com.ardaslegends.service.exceptions.claimbuild.ClaimBuildServiceException;
-import com.ardaslegends.service.exceptions.rpchar.RpCharServiceException;
+import com.ardaslegends.service.exceptions.logic.rpchar.RpCharServiceException;
 import com.ardaslegends.service.utils.ServiceUtils;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
