+++ src/main/java/com/ardaslegends/service/applications/RoleplayApplicationService.java
@@ -10,9 +10,9 @@ import com.ardaslegends.repository.applications.RoleplayApplicationRepository;
 import com.ardaslegends.service.AbstractService;
 import com.ardaslegends.service.dto.applications.CreateRpApplicatonDto;
 import com.ardaslegends.service.dto.applications.ApplicationVoteDto;
-import com.ardaslegends.service.exceptions.FactionServiceException;
-import com.ardaslegends.service.exceptions.PlayerServiceException;
-import com.ardaslegends.service.exceptions.applications.RoleplayApplicationServiceException;
+import com.ardaslegends.service.exceptions.logic.faction.FactionServiceException;
+import com.ardaslegends.service.exceptions.logic.player.PlayerServiceException;
+import com.ardaslegends.service.exceptions.logic.applications.RoleplayApplicationServiceException;
 import com.ardaslegends.service.utils.ServiceUtils;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
