+++ src/main/java/com/ardaslegends/service/exceptions/logic/applications/RoleplayApplicationServiceException.java
@@ -1,4 +1,4 @@
-package com.ardaslegends.service.exceptions.applications;
+package com.ardaslegends.service.exceptions.logic.applications;
 
 import com.ardaslegends.service.exceptions.ServiceException;
