+++ src/main/java/com/ardaslegends/service/war/BattleService.java
@@ -12,15 +12,11 @@ import com.ardaslegends.service.ArmyService;
 import com.ardaslegends.service.ClaimBuildService;
 import com.ardaslegends.service.PlayerService;
 import com.ardaslegends.service.dto.war.CreateBattleDto;
-import com.ardaslegends.service.exceptions.BattleServiceException;
-import com.ardaslegends.service.exceptions.PlayerServiceException;
-import com.ardaslegends.service.exceptions.WarServiceException;
-import com.ardaslegends.service.exceptions.army.ArmyServiceException;
-import com.ardaslegends.service.exceptions.claimbuild.ClaimBuildServiceException;
+import com.ardaslegends.service.exceptions.logic.war.BattleServiceException;
+import com.ardaslegends.service.exceptions.logic.army.ArmyServiceException;
 import com.ardaslegends.service.utils.ServiceUtils;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.javacord.api.DiscordApi;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
