+++ src/main/java/com/ardaslegends/service/war/WarService.java
@@ -8,9 +8,9 @@ import com.ardaslegends.repository.player.PlayerRepository;
 import com.ardaslegends.repository.WarRepository;
 import com.ardaslegends.service.AbstractService;
 import com.ardaslegends.service.dto.war.CreateWarDto;
-import com.ardaslegends.service.exceptions.FactionServiceException;
-import com.ardaslegends.service.exceptions.PlayerServiceException;
-import com.ardaslegends.service.exceptions.WarServiceException;
+import com.ardaslegends.service.exceptions.logic.faction.FactionServiceException;
+import com.ardaslegends.service.exceptions.logic.player.PlayerServiceException;
+import com.ardaslegends.service.exceptions.logic.war.WarServiceException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.javacord.api.DiscordApi;
