+++ src/test/java/com/ardaslegends/domain/FactionTest.java
@@ -1,6 +1,6 @@
 package com.ardaslegends.domain;
 
-import com.ardaslegends.service.exceptions.FactionServiceException;
+import com.ardaslegends.service.exceptions.logic.faction.FactionServiceException;
 import lombok.extern.slf4j.Slf4j;
 import org.junit.jupiter.api.Test;
