+++ src/test/java/com/ardaslegends/service/MovementServiceTest.java
@@ -8,10 +8,10 @@ import com.ardaslegends.repository.region.RegionRepository;
 import com.ardaslegends.service.dto.army.MoveArmyDto;
 import com.ardaslegends.service.dto.player.DiscordIdDto;
 import com.ardaslegends.service.dto.player.rpchar.MoveRpCharDto;
-import com.ardaslegends.service.exceptions.PlayerServiceException;
+import com.ardaslegends.service.exceptions.logic.player.PlayerServiceException;
 import com.ardaslegends.service.exceptions.ServiceException;
-import com.ardaslegends.service.exceptions.army.ArmyServiceException;
-import com.ardaslegends.service.exceptions.movement.MovementServiceException;
+import com.ardaslegends.service.exceptions.logic.army.ArmyServiceException;
+import com.ardaslegends.service.exceptions.logic.movement.MovementServiceException;
 import com.ardaslegends.service.utils.ServiceUtils;
 import lombok.extern.slf4j.Slf4j;
 import org.junit.jupiter.api.BeforeEach;
