+++ src/test/java/com/ardaslegends/service/PathfinderTest.java
@@ -2,7 +2,7 @@ package com.ardaslegends.service;
 
 import com.ardaslegends.domain.*;
 import com.ardaslegends.repository.region.RegionRepository;
-import com.ardaslegends.service.exceptions.PathfinderServiceException;
+import com.ardaslegends.service.exceptions.logic.movement.PathfinderServiceException;
 import com.ardaslegends.service.war.WarService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
